// apps/web/src/lib/api/generated/models/reviewRunDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewRunStateDto } from './reviewRunStateDto.ts';

export interface ReviewRunDto {
	created_at: string;
	definition: string;
	/** @nullable */
	finished_at?: string | null;
	id: string;
	origin: unknown;
	project_id: string;
	/** @nullable */
	started_at?: string | null;
	state: ReviewRunStateDto;
	subject: unknown;
}